requirements.txt
pandas~=1.4.1
numpy~=1.22.2
python-binance~=1.0.15
TA-Lib~=0.4.24
pandas-ta~=0.3.14b0
jupyter~=1.0.0

